+++ packages/utils/src/obj/index.ts
@@ -1,0 +1,14 @@
+import { entriesOf } from './entries-of';
+import { keysOf } from './keys-of';
+
+/**
+ * Utility functions for objects.
+ *
+ * @since 0.0.8
+ */
+export const Obj = {
+    entries: entriesOf,
+    keys: keysOf,
+};
+
+export { entriesOf, keysOf };